// constants.d.ts
export declare const RESUME_DATA: {
    personalInfo: {
        name: string;
        title: string;
        email: string;
        phone: string;
        location: string;
        profilePic: string;
        social: {
            linkedin: string;
            github: string;
            portfolio: string;
            resume: string;
        };
        summary: string;
    };
    experience: {
        company: string;
        role: string;
        location: string;
        period: string;
        tech: string[];
        description: string[];
    }[];
    projects: {
        name: string;
        tech: string[];
        description: string;
        demoUrl: string;
        repoUrl: string;
    }[];
    skills: {
        name: string;
        icon: string;
    }[];
    education: {
        degree: string;
        institution: string;
        period: string;
        coursework: string;
    }[];
    responsibility: {
        role: string;
        period: string;
        description: string;
    }[];
    certifications: string[];
    spokenLanguages: string[];
};
export declare const SERVICES_DATA: {
    title: string;
    description: string;
    icon: string;
}[];
export declare const MY_EDGE_DATA: {
    title: string;
    description: string;
    icon: string;
}[];
export declare const PROCESS_DATA: {
    step: string;
    title: string;
    description: string;
    icon: string;
    tags: string[];
}[];
export declare const NAV_LINKS: {
    label: string;
    href: string;
}[];
//# sourceMappingURL=constants.d.ts.map